Move to resource.go and improve comments
