--- service/controller/app/resource/authtokenmigration/delete.go
@@ -1,8 +1,0 @@
-package authtokenmigration
-
-import "context"
-
-func (r Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	// no-op
-	return nil
-}
